Separate application Rust test suites
Move API, embedded CLI, and file-service tests out of production modules while preserving their module scope.
